Support for Scout shouldBeSearchable
Scout's `shouldBeSearchable` function fires a delete record on all save/create events. This makes sense as Scout does not keep track if models have been added before or not. 

This update checks for the model's existence before attempting to delete it.

+++ src/Typesense.php
@@ -123,9 +123,14 @@ class Typesense
         /**
          * @var $document Document
          */
-        $document = $collectionIndex->getDocuments()[(string) $modelId];
+        $document = $collectionIndex->getDocuments()[(string) $modelId] ?? null;
 
-        return $document->delete();
+        // dd($document);
+        if($document) {
+            return $document->delete();
+        }
+
+        return [];
     }
 
     /**
